Create Component for Blob Properties
